ZNTA-2215 - move classes to reuse
