+++ api/zanata-common-api/src/main/java/org/zanata/common/IssuePriority.java
@@ -18,7 +18,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.zanata.model;
+package org.zanata.common;
 
 public enum IssuePriority {
     Critical, Major, Minor
