+++ server/gwt-shared/src/main/java/org/zanata/webtrans/shared/rest/dto/TransReviewCriteria.java
@@ -18,10 +18,10 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.zanata.rest.review.dto;
+package org.zanata.webtrans.shared.rest.dto;
 
 import org.codehaus.jackson.annotate.JsonIgnoreProperties;
-import org.zanata.model.IssuePriority;
+import org.zanata.common.IssuePriority;
 
 /**
  * @author Patrick Huang <a href="mailto:pahuang@redhat.com">pahuang@redhat.com</a>
