+++ server/zanata-model/src/main/java/org/zanata/model/ReviewCriteria.java
@@ -30,6 +30,8 @@ import javax.persistence.Enumerated;
 import javax.persistence.NamedQueries;
 import javax.persistence.NamedQuery;
 
+import org.zanata.common.IssuePriority;
+
 
 /**
  * @author Patrick Huang <a href="mailto:pahuang@redhat.com">pahuang@redhat.com</a>
